ai-service/requirements.txt
# LangChain and related packages
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-google-genai>=0.0.5

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# PDF processing
pypdf>=3.17.1

# Environment variables
python-dotenv>=1.0.0

# Google Generative AI SDK (provides `google.generativeai` import)
google-generative-ai>=0.4.0

# Other utilities
typing-extensions>=4.8.0
requests>=2.31.0
numpy>=1.24.0
tqdm>=4.66.1

# Optional but recommended
jupyter>=1.0.0  # for notebook support
pandas>=2.1.0   # for data handling
